+++ packages/components/grid/src/grid.ts
@@ -1,22 +1,24 @@
 import type { CSSResultGroup, PropertyValues, TemplateResult } from 'lit';
 import type { GridSorter, GridSorterChange } from './sorter.js';
 import type { GridFilter, GridFilterChange } from './filter.js';
+import type { GridColumnGroup } from './column-group.js';
 import type { Virtualizer } from '@lit-labs/virtualizer/Virtualizer.js';
 import type { ScopedElementsMap } from '@open-wc/scoped-elements/lit-element.js';
 import type { DataSource, EventEmitter } from '@sl-design-system/shared';
 import { localized } from '@lit/localize';
 import { virtualize, virtualizerRef } from '@lit-labs/virtualizer/virtualize.js';
 import { ScopedElementsMixin } from '@open-wc/scoped-elements/lit-element.js';
-import { ArrayDataSource, SelectionController, event, isSafari } from '@sl-design-system/shared';
+import { ArrayDataSource, SelectionController, event, getValueByPath, isSafari } from '@sl-design-system/shared';
 import { LitElement, html } from 'lit';
 import { property, query, state } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
+import { GridViewModel, GridViewModelGroup } from './view-model.js';
 import { GridColumn } from './column.js';
-import { GridColumnGroup } from './column-group.js';
 import styles from './grid.scss.js';
 import { GridSelectionColumn } from './selection-column.js';
 import { GridActiveItemChangeEvent, GridEvent, GridItemDropEvent, GridItemEvent } from './events.js';
 import { GridFilterColumn } from './filter-column.js';
+import { GridGroupHeader } from './group-header.js';
 import { GridSortColumn } from './sort-column.js';
 
 export type GridItemParts<T> = (model: T) => string | undefined;
@@ -37,7 +39,9 @@ export type GridDropFilter<T> = (item: T) => boolean | 'between' | 'on-top';
 export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
   /** @private */
   static get scopedElements(): ScopedElementsMap {
-    return {};
+    return {
+      'sl-grid-group-header': GridGroupHeader
+    };
   }
 
   /** @private */
@@ -95,9 +99,6 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
   /** Emits when the active item changes */
   @event() activeItemChange!: EventEmitter<GridActiveItemChangeEvent<T>>;
 
-  /** The columns in the grid. */
-  @state() columns: Array<GridColumn<T>> = [];
-
   /** Provide your own implementation for getting the data. */
   @property({ attribute: false }) dataSource?: DataSource<T>;
 
@@ -133,8 +134,14 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
   /** An array of items to be displayed in the grid. */
   @property({ type: Array }) items?: T[];
 
+  /** The path to the attribute to group the items on. */
+  @property({ reflect: true, attribute: 'items-group-by' }) itemsGroupBy?: string;
+
   /** Custom parts to be set on the `<tr>` so it can be styled externally. */
   @property({ attribute: false }) itemParts?: GridItemParts<T>;
+
+  /** The model used for rendering the grid. */
+  @property({ attribute: false }) model = new GridViewModel<T>(this);
 
   /** Hide the border around the grid when true. */
   @property({ type: Boolean, reflect: true, attribute: 'no-border' }) noBorder?: boolean;
@@ -193,12 +200,16 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
 
   override willUpdate(changes: PropertyValues<this>): void {
     if (changes.has('items')) {
-      if (this.items) {
-        this.dataSource = new ArrayDataSource(this.items);
-        this.selection.size = this.dataSource.size;
+      this.dataSource = this.items ? new ArrayDataSource(this.items) : undefined;
+      this.model.dataSource = this.dataSource;
+      this.selection.size = this.dataSource?.size ?? 0;
+    }
+
+    if (changes.has('itemsGroupBy')) {
+      if (this.itemsGroupBy) {
+        this.dataSource?.setGroupBy(this.itemsGroupBy);
       } else {
-        this.dataSource = undefined;
-        this.selection.size = 0;
+        this.dataSource?.removeGroupBy();
       }
     }
   }
@@ -207,10 +218,8 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
     super.updated(changes);
 
     if (changes.has('dataSource')) {
-      this.dataSource?.addEventListener('sl-update', () => {
-        this.selection.size = this.dataSource?.size ?? 0;
-        this.requestUpdate();
-      });
+      this.model.dataSource = this.dataSource;
+      this.selection.size = this.dataSource?.size ?? 0;
 
       this.#applyFilters();
       this.#applySorters();
@@ -235,7 +244,7 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
         </thead>
         <tbody @visibilityChanged=${this.#onVisibilityChanged} part="tbody">
           ${virtualize({
-            items: this.dataSource?.filteredItems,
+            items: this.model.rows,
             renderItem: (item, index) => this.renderItem(item, index)
           })}
         </tbody>
@@ -244,7 +253,7 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
   }
 
   renderStyles(): TemplateResult {
-    const rows = this.#getHeaderRows(this.columns);
+    const rows = this.model.headerRows;
 
     return html`
       ${rows.slice(0, -1).map((row, rowIndex) => {
@@ -268,7 +277,7 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
             ${
               col.sticky
                 ? `
-                  inset-inline-start: ${this.#getStickyColumnOffset(index)}px;
+                  inset-inline-start: ${this.model.getStickyColumnOffset(index)}px;
                   position: sticky;
                 `
                 : ''
@@ -281,7 +290,7 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
   }
 
   renderHeader(): TemplateResult {
-    const rows = this.#getHeaderRows(this.columns),
+    const rows = this.model.headerRows,
       showSelectionHeader =
         this.selection.size > 0 &&
         (this.selection.areSomeSelected() || this.selection.areAllSelected()) &&
@@ -310,7 +319,11 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
   }
 
   renderItem(item: T, index: number): TemplateResult {
-    const rows = this.#getHeaderRows(this.columns),
+    return item instanceof GridViewModelGroup ? this.renderGroupRow(item) : this.renderItemRow(item, index);
+  }
+
+  renderItemRow(item: T, index: number): TemplateResult {
+    const rows = this.model.headerRows,
       selected = this.selection.isSelected(item),
       parts = [
         'row',
@@ -336,17 +349,38 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
     `;
   }
 
+  renderGroupRow(group: GridViewModelGroup): TemplateResult {
+    const expanded = this.model.getGroupState(group.value),
+      selectable = !!this.model.columns.find(col => col instanceof GridSelectionColumn),
+      selected = this.model.getGroupSelection(group.value);
+
+    return html`
+      <tr part="group">
+        <td part="group-header">
+          <sl-grid-group-header
+            @sl-select=${(event: CustomEvent<boolean>) => this.#onGroupSelect(event, group)}
+            @sl-toggle=${(event: CustomEvent<boolean>) => this.#onGroupToggle(event, group)}
+            .expanded=${expanded}
+            .heading=${group.value}
+            .selectable=${selectable}
+            .selected=${selected}
+          ></sl-grid-group-header>
+        </td>
+      </tr>
+    `;
+  }
+
   /** Updates the `width` of all columns which have `autoWidth` set to `true`. */
   async recalculateColumnWidths(): Promise<void> {
     // Do not remove, this is needed; not sure why
     await this.updateComplete;
 
-    const rows = this.#getHeaderRows(this.columns);
+    const rows = this.model.headerRows;
 
     rows[rows.length - 1]
       .filter(col => !col.hidden && col.autoWidth)
       .forEach(col => {
-        const index = this.columns.indexOf(col),
+        const index = this.model.columns.indexOf(col),
           cells = this.renderRoot.querySelectorAll<HTMLElement>(`:where(td, th):nth-child(${index + 1})`);
 
         col.width = Array.from(cells).reduce((acc, cur) => {
@@ -364,10 +398,10 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
     // Since we set an explicit width for the `<thead>` and `<tbody>`, we also need
     // to set an explicit width for all the `<tr>` elements. Otherwise, the sticky columns
     // will not be sticky when you scroll horizontally.
-    const rowWidth = this.columns.reduce((acc, cur) => acc + Number(cur?.width ?? 0), 0);
+    const rowWidth = this.model.columns.reduce((acc, cur) => acc + Number(cur?.width ?? 0), 0);
     this.style.setProperty('--sl-grid-row-width', `${rowWidth}px`);
 
-    this.requestUpdate('columns');
+    this.requestUpdate();
   }
 
   #onClickRow(event: Event, item: T): void {
@@ -544,7 +578,22 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
     this.#applyFilters();
   }
 
-  #onSlotchange(event: Event & { target: HTMLSlotElement }): void {
+  #onGroupSelect(event: CustomEvent<boolean>, group: GridViewModelGroup): void {
+    const items = this.dataSource?.filteredItems ?? [],
+      groupItems = items.filter(item => getValueByPath(item, this.itemsGroupBy) === group.value);
+
+    if (event.detail) {
+      groupItems.forEach(item => this.selection.select(item));
+    } else {
+      groupItems.forEach(item => this.selection.deselect(item));
+    }
+  }
+
+  #onGroupToggle(event: CustomEvent<boolean>, group: GridViewModelGroup): void {
+    this.model.toggleGroup(group.value, event.detail);
+  }
+
+  async #onSlotchange(event: Event & { target: HTMLSlotElement }): Promise<void> {
     const elements = event.target.assignedElements({ flatten: true }),
       columns = elements.filter((el): el is GridColumn<T> => el instanceof GridColumn);
 
@@ -570,7 +619,11 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
       }
     });
 
-    this.columns = columns;
+    // Wait for all the columns to update; the column group especially
+    // needs time for the slotchange event to fire.
+    await Promise.allSettled(columns.map(async col => col.updateComplete));
+
+    this.model.columnDefinitions = columns;
   }
 
   #onSortDirectionChange({ target }: Event & { target: GridSorter<T> }): void {
@@ -644,24 +697,5 @@ export class Grid<T = any> extends ScopedElementsMixin(LitElement) {
     this.dataSource.update();
     this.requestUpdate();
     this.gridStateChange.emit(new GridEvent('sl-grid-state-change', this));
-  }
-
-  #getHeaderRows(columns: Array<GridColumn<T>>): Array<Array<GridColumn<T>>> {
-    const children = columns
-      .filter((col): col is GridColumnGroup<T> => col instanceof GridColumnGroup)
-      .reduce((acc: Array<Array<GridColumn<T>>>, cur) => {
-        return [...acc, ...this.#getHeaderRows(cur.columns)];
-      }, []);
-
-    return children.length ? [[...columns], children.flat(2)] : [[...columns]];
-  }
-
-  #getStickyColumnOffset(index: number): number {
-    return this.columns
-      .slice(0, index)
-      .filter(col => !col.hidden)
-      .reduce((acc, { width = 0 }) => {
-        return acc + width;
-      }, 0);
   }
 }