+++ packages/components/tooltip/src/tooltip.ts
@@ -1,28 +1,69 @@
 import type { CSSResultGroup, PropertyValues, TemplateResult } from 'lit';
 import type { PopoverPosition } from '@sl-design-system/shared';
-import { AnchorController, EventsController, popoverPolyfillStyles } from '@sl-design-system/shared';
+import { AnchorController, EventsController } from '@sl-design-system/shared';
 import { LitElement, html } from 'lit';
 import { property } from 'lit/decorators.js';
 import styles from './tooltip.scss.js';
+
+// Workaround for missing type in @open-wc/scoped-elements
+declare global {
+  interface ShadowRoot {
+    createElement(tagName: string): HTMLElement;
+  }
+}
+
+export interface TooltipOptions {
+  /**
+   * This determines the context that is used to create the `<sl-tooltip>` element. If
+   * not provided, the tooltip will be created on the document. But that assumes that
+   * the tooltip web component is already defined in the document. Otherwise you can
+   * specify a shadow root that has the tooltip defined in it.
+   */
+  context?: Document | ShadowRoot;
+
+  /**
+   * This is the node where the tooltip will be added to. This can be useful when
+   * you don't want the tooltip to be added next to the anchor element. If not provided,
+   * it will be added next to the anchor element.
+   */
+  parentNode?: Node;
+}
 
 let nextUniqueId = 0;
 
 /**
  * Tooltip component.
+ *
+ * @slot default - The slot for the tooltip content.
  */
 export class Tooltip extends LitElement {
-  /** @private */
-  static override styles: CSSResultGroup = [popoverPolyfillStyles, styles];
+  /** The default padding of the arrow. */
+  static arrowPadding = 16;
 
-  static lazy(target: Element, callback: (target: Tooltip) => void): void {
+  /** The default offset of the tooltip to its anchor. */
+  static offset = 12;
+
+  /** @private */
+  static override styles: CSSResultGroup = styles;
+
+  /** The default margin between the tooltip and the viewport. */
+  static viewportMargin = 8;
+
+  /** To attach the `sl-tooltip` to the DOM tree and anchor element */
+  static lazy(target: Element, callback: (target: Tooltip) => void, options: TooltipOptions = {}): void {
     const createTooltip = (): void => {
       // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
-      const tooltip = document.createElement('sl-tooltip') as Tooltip;
+      const tooltip = (options.context ?? document).createElement('sl-tooltip') as Tooltip;
       tooltip.id = `sl-tooltip-${nextUniqueId++}`;
 
       callback(tooltip);
 
-      target.parentNode?.insertBefore(tooltip, target.nextSibling);
+      if (options.parentNode) {
+        options.parentNode.appendChild(tooltip);
+      } else {
+        target.parentNode?.insertBefore(tooltip, target.nextSibling);
+      }
+
       target.setAttribute('aria-describedby', tooltip.id);
 
       tooltip.anchorElement = target as HTMLElement;
@@ -35,11 +76,13 @@ export class Tooltip extends LitElement {
     ['focusin', 'pointerover'].forEach(eventName => target.addEventListener(eventName, createTooltip));
   }
 
-  /** Tooltip max-width. */
-  @property({ type: Number, attribute: 'max-width' }) maxWidth?: number;
-
   /** Controller for managing anchoring. */
-  #anchor = new AnchorController(this, { maxWidth: this.maxWidth });
+  #anchor = new AnchorController(this, {
+    arrowElement: '.arrow',
+    arrowPadding: Tooltip.arrowPadding,
+    offset: Tooltip.offset,
+    viewportMargin: Tooltip.viewportMargin
+  });
 
   /** Events controller. */
   #events = new EventsController(this);
@@ -50,7 +93,6 @@ export class Tooltip extends LitElement {
 
   #onHide = ({ target }: Event): void => {
     if (this.#matchesAnchor(target as Element)) {
-      this.anchorElement = undefined;
       this.hidePopover();
     }
   };
@@ -62,7 +104,13 @@ export class Tooltip extends LitElement {
     }
   };
 
-  /** Tooltip position. */
+  /** The maximum width of the tooltip. */
+  @property({ type: Number, attribute: 'max-width' }) maxWidth?: number;
+
+  /**
+   * Position of the tooltip relative to its anchor.
+   * @type {'top' | 'right' | 'bottom' | 'left' | 'top-start' | 'top-end' | 'right-start' | 'right-end' | 'bottom-start' | 'bottom-end' | 'left-start' | 'left-end'}
+   */
   @property() position: PopoverPosition = 'top';
 
   override connectedCallback(): void {
@@ -81,6 +129,8 @@ export class Tooltip extends LitElement {
   }
 
   override willUpdate(changes: PropertyValues<this>): void {
+    super.willUpdate(changes);
+
     if (changes.has('maxWidth')) {
       this.#anchor.maxWidth = this.maxWidth;
     }
@@ -91,13 +141,15 @@ export class Tooltip extends LitElement {
   }
 
   override render(): TemplateResult {
-    return html` <slot></slot>
+    return html`
+      <slot></slot>
       <div class="arrow">
         <svg xmlns="http://www.w3.org/2000/svg" width="16" height="6">
           <path
             d="M14.48 5.411c.384.374.938.588 1.52.589H0c.582-.001 1.136-.215 1.52-.589L6.475.59c.807-.785 2.241-.785 3.048 0L14.48 5.41Z"
           />
         </svg>
-      </div>`;
+      </div>
+    `;
   }
 }