+++ packages/components/grid/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @sl-design-system/grid
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`dbc032b`](https://github.com/sl-design-system/components/commit/dbc032b3a7587dfcbdb6a2118330b039765cf0fb), [`5da216b`](https://github.com/sl-design-system/components/commit/5da216b3713c328eba06113d77d642462e1f05fc)]:
+  - @sl-design-system/shared@0.2.5
+  - @sl-design-system/popover@0.1.5
+  - @sl-design-system/checkbox@0.0.22
+  - @sl-design-system/text-field@0.1.17
+
+## 0.1.0
+
+### Minor Changes
+
+- [#987](https://github.com/sl-design-system/components/pull/987) [`8ce8ec6`](https://github.com/sl-design-system/components/commit/8ce8ec64aafbc94916adfe7b7c25968a1b0975db) - Fix grid selection with DataSource
+
+### Patch Changes
+
+- Updated dependencies [[`216d62e`](https://github.com/sl-design-system/components/commit/216d62eb5a16277b4ea2767ea0530e570bf40abf), [`216d62e`](https://github.com/sl-design-system/components/commit/216d62eb5a16277b4ea2767ea0530e570bf40abf)]:
+  - @sl-design-system/popover@0.1.4
+  - @sl-design-system/shared@0.2.4
+  - @sl-design-system/checkbox@0.0.21
+  - @sl-design-system/text-field@0.1.16
+
+## 0.0.36
+
+### Patch Changes
+
+- [#939](https://github.com/sl-design-system/components/pull/939) [`afc3d60`](https://github.com/sl-design-system/components/commit/afc3d606c20409b4ad2d589ffc0b899d3f853997) - Add support for grouping rows by an item property
+
+- Updated dependencies [[`afc3d60`](https://github.com/sl-design-system/components/commit/afc3d606c20409b4ad2d589ffc0b899d3f853997)]:
+  - @sl-design-system/shared@0.2.3
+  - @sl-design-system/checkbox@0.0.20
+  - @sl-design-system/popover@0.1.3
+  - @sl-design-system/text-field@0.1.15
+
 ## 0.0.35
 
 ### Patch Changes