+++ packages/themes/sanoma-learning/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sanomalearning/slds-sanoma-learning
 
+## 0.3.0
+
+### Minor Changes
+
+- [#990](https://github.com/sl-design-system/components/pull/990) [`89d6d7f`](https://github.com/sl-design-system/components/commit/89d6d7f7d1686cab7c77d0f6cdf18be1367f5edc) - Added accordion tokens
+
+## 0.2.4
+
+### Patch Changes
+
+- [#985](https://github.com/sl-design-system/components/pull/985) [`9c88751`](https://github.com/sl-design-system/components/commit/9c887516bacb0a335d8d2283e47d9a78f359a334) - Added tokens for Tab component
+
 ## 0.2.3
 
 ### Patch Changes